move locale parsing to different file and write config parser
