+++ Docs/src/data/.gitignore
@@ -1,0 +1,2 @@
+*
+!.gitignore
